clang-tidy : corrections manuelles des alertes Code scanning restantes
- Complexité cognitive : fonctions découpées en étapes nommées (main, LevelWriter::buildJson,
  CharacterSheet, Dialogue, EnemyAi, MainWindow::buildUi, GameViewport::tick…).
- Défauts réels : cast élargissant après la multiplication (AssetMarker, LineOfSight), arrondi
  par std::lround (CharacterSheetValues), compteurs flottants de PixelCanvas, garde nullptr dans
  MainWindow::raisePanel, cas Key::E manquant dans KeyName.
- Champs d'agrégats explicites, std::array, unions DirectXMath lues par l'API, qobject_cast,
  constantes inutilisées supprimées, nommage aligné sur .clang-tidy.
- JADG_ASSERT n'enveloppe plus la condition dans une négation.
- 9 NOLINT justifiés : drapeaux MINIDUMP_TYPE, balayage des codes de touche, méthodes lues par
  QML, reinterpret_cast imposé par QImage.

+++ Source/Core/Resources/AssetMarker.cpp
@@ -104,7 +104,8 @@ MarkerImage assetMarker(std::string_view key, int width, int height) {
                 std::fabs(static_cast<double>(y) - attendueMontante) <= epaisseur;
             // Deux diagonales barrent la vignette : un marqueur doit se voir COMME un marqueur, et
             // ne jamais passer pour une illustration definitive livree un peu vite.
-            image.pixels[static_cast<std::size_t>((y * width) + x)] = surDiagonale ? trait : fond;
+            image.pixels[(static_cast<std::size_t>(y) * static_cast<std::size_t>(width)) +
+                         static_cast<std::size_t>(x)] = surDiagonale ? trait : fond;
         }
     }
     return image;
